♻️ フェーズ4: データ変換ロジックの最適化
- useDataTransformフックを作成してデータ変換処理を独立化
- seriesedBooks計算ロジック、統計処理を専用フックに移動
- satisfiesを使用して型安全性を向上
- useBookDataからデータ変換ロジックを分離し単一責任を実現
